Uses the LoadPlugin syntax with bracket when supported
This is fully backwards compatible.

The version check is done exactly how it's supposed to ; I checked the
sources of collectd to be sure there's no conflict.

Globals has always defaulted to false so I'm making it default to false
in this module. I'm not setting it undef because that's a boolean and we
got issues with checking if a boolean is false or undef (see #182).