Fix collectd::plugin::write_http for RedHat
This change reverts commit a334b6e and closes issue #251. The main RPM
repositories (eg, Fedora and EPEL) don't ship the collectd-write_http
package (instead the plugin is included in the core collectd package).
In case the collectd-write_http needs to be installed, it can be done
outside of the collectd manifests.