plugin/ping: Transition from defined type to class
This brings the ping plugin inline with the rest of the plugins.
It transitions it from a defined type that could be called multiple
times creating multiple `<Plugin ping></Plugin>`, definitions which
appears not to be valid, to a class.

As one can pass in an array to hosts you can still ping multiple
targets.