Add OpenBSD switch on wordexp
This work is required to support the OpenBSD operating system due to its
lack of the "wordexp" function.  As noted in the collectd.conf(5) man
page, expansion of the shell-like wildcards depend on the "wordexp".  As
such, this work adds a $has_wordexp param to skip this kind of
sub-config loading on OpenBSD.