Remove interval parameter from write plugins.
It makes no sense to configure interval for write plugins.
Collectd has no concept of interval for write plugins and if
configured, it is silently ignored. It is therefore better to remove
it from the configuration so that manifests don't give a false
impression that this parameter does something useful.