Use one operating system for most RSpec tests
The majority of Puppet classes and defined types in this module do not
have behaviour which changes depending on the operating system. Mock one
operating system for most tests, except for collectd_init_spec when
mocking multiple operating systems makes sense.

This attempts to address #633 , and should reduce test runtimes
considerably with little change in the confidence of tests actually
picking up problems.