Use different file name for processes plugin config
This change fixes collectd::plugin::processes to work on RHEL-based
systems again after a conflicting change on puppet-collectd v10.0.1.

It uses a different filename for the processes plugin configuration
in a way that it will no longer overwrite a vendor-supplied file that
is being wiped as part of collectd::plugin.