Add CentOS 8 support
The main differences for CentOS 8 are:
* There is now a new collectd-utils package containing collectdctl
* collectd-python is now a seperate sub package
* python is python3.
* pip command is now pip3
* The rabbitmq plugin no longer works due to lack of python3 support

The python_dir fact will fall back to `/usr/libexec/platform-python`
if `python3` does not exist on the path. This is the python
that collectd is built against and significantly requires.

Note that setting `collectd::ci_package_repo` true will currently
result in yum being configured with a non existant repo since
https://pkg.ci.collectd.org/rpm/ does not exist
for 8.

https://github.com/collectd/collectd-ci/issues/35