Add backwards compatibility in transition to Stdlib::Host
Patch [1] changed all `port` parameters to use `Stdlib::Port`. When
moving from String type to the new type, this patch is backwards
incompatible leading to type errors when applying manifest. This patch
is using Variant type to make this change backwards compatible and ease
transition to new versions of puppet-collectd.

[1] https://github.com/voxpupuli/puppet-collectd/commit/d7b79c502bd10587d7498ddc1d19484eabfd7380