Allow the getval task to work with customized hostnames
The default hostname used by collectd is system dependent as retuned by
gethostname(3) (On FreeBSD it's the FQDN, while on Linux it's the
non-FQDN).  The python method socket.gethostname() always return a
non-FQDN, and the user has the ability to tune the hostname as he wish
in the collectd configuration (hostname parameter).

This change allow to use a customized hostname by using an optional
argument "hostname".  When not provided, the non-FQDN and the FQDN are
tried in turn to fetch the requested metric value.