MySQL database plugin: allow no password; default Host, User, and Password to undef (#970)
* Allow empty password

Both MySQL and MariaDB now support authentication types that do not
require a password to be set, eg. socket authentication, which is
now the default at least on Debian-based distributions on new
installations.

Currently the module hardcodes 'UNSET' for the password, this
trivial change makes supplying the passsword optional.

* make host default to 'localhost'

Using 'localhost' as the default host to connect to better matches
what the clients for these databases default to.

* make username optional

The collectd mysql plugin documentation has examples where a socket
is configured for accessing the database, similar to the default
behaviour of mysqlclient. Along with the previous change to set the
default hostname to localhost, this matches up better with what
users of these tools expect.

* fix test

adjust to reflect changes introduced by previous commits

* add length restriction and better type

use Stdlib::Host for $hostname and make $username and $password
require a non-empty string