Remove unreliable acceptance tests
These tests are very fragile: they depend on the version of collectd,
the version of python bundled with collectd, the version of python
available on the system.  Sometimes these version match, more often they
don't.  These tests also depend on external projects which are not
actively maintained anymore and we face race conditions hard to debug
that make the CI fail.

Since the aspect we are already interested in are already tested with
puppet-rspec, drop these test which depends on too much external things.