Be explicit for requires of collectd.d
Previously all the concat's created typically in
`/etc/collect.d` were subject to

```
File['collectd.d'] -> Concat <| tag == 'collectd' |>
```
to ensure the directory was in place first.

This turns out to be too greedy and can cause problems
as per #1000.

Instead be explicit about requiring the directory on
the concat files that are bing created.

Consequence of this if concat files are being created
in `/etc/collect.d` outside this module they will now
have to require this directory. At worst such configuration
would require a second puppet run.