Rely on autorequires of files when using concat
Previously all the concat's created typically in /etc/collect.d were subject to

File['collectd.d'] -> Concat <| tag == 'collectd' |>
to ensure the directory was in place first.

This turns out to be too greedy and can cause problems as per #1000.

This declaration is unnecessary as concat creates a file type
and this auto-requires the directory one level above.